Extract FocusTrap to its own file
